+++ python/CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Improve golang types
+
 ## [python/v0.5.0] — 2026-05-07
 
 ### Added
@@ -47,6 +53,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Changelog generation
 
+[unreleased]: https://github.com/mantyx-io/mantyx-sdk/compare/python/v0.5.0..HEAD
 [python/v0.5.0]: https://github.com/mantyx-io/mantyx-sdk/compare/python/v0.4.1..python/v0.5.0
 [python/v0.4.1]: https://github.com/mantyx-io/mantyx-sdk/compare/python/v0.4.0..python/v0.4.1
 [python/v0.4.0]: https://github.com/mantyx-io/mantyx-sdk/compare/python/v0.3.0..python/v0.4.0